# Blocks: determine block names from filename convention instead of disk access (2)