--- a/projects/plugins/jetpack/changelog/add-get_block_name_from_path_convention
+++ b/projects/plugins/jetpack/changelog/add-get_block_name_from_path_convention
@@ -1,0 +1,5 @@
+Significance: patch
+Type: other
+Comment: Adds a unit test for a change in the blocks repo
+
+