## Proposed changes:

* When registering blocks, use the composition of the path instead of reading the block.json file to determine the filename.

For example, when we see the path `extensions/blocks/ai-assistant`, we know the block name will be `jetpack/ai-assistant`. This holds true 99% of the time.

See also: pdWQjU-RU-p2#comment-932

### Difference from last PR:

* I wasn't able to deploy the last PR, #39307 because of E2E test failures. 
* They were caused by the removal of `$block_type = self::get_path_to_block_metadata( $slug );` which uses `JETPACK__PLUGIN_FILE` to change paths on some systems. 
* Fixed in 9aa50cd8a9e036ff1e6c9acae7d92fd71777ef70

### Why - Performance

This is a performance based change, saving us from having to `json_decode( file_get_contents( 'block.json' ) )` to get the block name. On wpcom, here are some example trials when wrapping `wp_json_file_decode()` in `ac_start()` and `ac_stop()` ([defined here](https://github.com/mreishus/wp-misc/blob/master/debug-helpers.php#L90-L125)).

**Before**:
```
[1            b3bf7] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=hpkmx1yd
[1            b3bf7] AC Debug: 'decode' ran 191 times, total time: 8.51 ms, avg time: 0.04 ms
[1            49f43] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=6gbhpgny
[1            49f43] AC Debug: 'decode' ran 191 times, total time: 8.68 ms, avg time: 0.05 ms
[1            32755] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=4e4ol5d3
[1            32755] AC Debug: 'decode' ran 191 times, total time: 7.44 ms, avg time: 0.04 ms
[1            73db4] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=b48rjlzy
[1            73db4] AC Debug: 'decode' ran 191 times, total time: 8.7 ms, avg time: 0.05 ms
[1            161e4] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=sq357how
[1            161e4] AC Debug: 'decode' ran 191 times, total time: 10.1 ms, avg time: 0.05 ms
[1            837d5] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=mooppp68
[1            837d5] AC Debug: 'decode' ran 191 times, total time: 7 ms, avg time: 0.04 ms
[1            774a1] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=uz4s5jjo
[1            774a1] AC Debug: 'decode' ran 191 times, total time: 7.55 ms, avg time: 0.04 ms
[1            886d6] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=p73evouz
[1            886d6] AC Debug: 'decode' ran 191 times, total time: 7.04 ms, avg time: 0.04 ms
[1            83988] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=c4lm2jmq
[1            83988] AC Debug: 'decode' ran 191 times, total time: 7.2 ms, avg time: 0.04 ms
[1            47300] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=wej26sn6
[1            47300] AC Debug: 'decode' ran 191 times, total time: 10.83 ms, avg time: 0.06 ms
```

**After**:
```
[1            a99ca] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=7ab0dpyx
[1            a99ca] AC Debug: 'decode' ran 108 times, total time: 5.34 ms, avg time: 0.05 ms
[1            f3127] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=1armlnw7
[1            f3127] AC Debug: 'decode' ran 108 times, total time: 4.83 ms, avg time: 0.04 ms
[1            f25d3] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=k8xd0e8k
[1            f25d3] AC Debug: 'decode' ran 108 times, total time: 5.4 ms, avg time: 0.05 ms
[1            4abbd] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=zwue94yv
[1            4abbd] AC Debug: 'decode' ran 108 times, total time: 4.09 ms, avg time: 0.04 ms
[1            bf358] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=38dpegr3
[1            bf358] AC Debug: 'decode' ran 108 times, total time: 4.06 ms, avg time: 0.04 ms
[1            75e77] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=17qwe8om
[1            75e77] AC Debug: 'decode' ran 108 times, total time: 3.98 ms, avg time: 0.04 ms
[1            afe3a] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=3smekv6c
[1            afe3a] AC Debug: 'decode' ran 108 times, total time: 4.03 ms, avg time: 0.04 ms
[1            2dc15] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=9tobu3yl
[1            2dc15] AC Debug: 'decode' ran 108 times, total time: 4.26 ms, avg time: 0.04 ms
[1            4bc37] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=ogsz24u0
[1            4bc37] AC Debug: 'decode' ran 108 times, total time: 6.31 ms, avg time: 0.06 ms
[1            18d97] https://public-api.wordpress.com/rest/v1.1/test/1?defeatcache=0hzbcwad
[1            18d97] AC Debug: 'decode' ran 108 times, total time: 5.01 ms, avg time: 0.05 ms
```

### Exceptions

* The path blocks/premium-content defines `premium-content\/container`, not `jetpack/premium-content`.

This is added to a `$breaks_convention` array, which means we fall through to the old method instead of looking at the path.

### Exceptions - Catching future ones

The unit test `test_get_block_name_from_path_convention_matches_get_block_name()` does the path-based method and the block.json method for all blocks, and makes sure they're the same. Therefore, if a new block which is added which breaks convention, the unit tests will fail in order to alert us.


### Other information:

- [ ] Have you written new tests for your changes, if applicable?
- [ ] Have you checked the E2E test CI results, and verified that your changes do not break them?
- [ ] Have you tested your changes on WordPress.com, if applicable (if so, you'll see a generated comment below with a script to run)?

## Jetpack product discussion
<!-- If you're an Automattician, include a shortlink to the p2 discussion with Jetpack Product here. -->
<!-- Make sure any changes to existing products have been discussed and agreed upon -->

## Does this pull request change what data or activity we track or use?
<!--- If so, please add the "[Status] Needs Privacy Updates" label and explain what changes there are. -->
<!--- Check existing Jetpack support documents for a preview of the information we need. -->

## Testing instructions:
<!-- If you were reviewing this PR, how would you like the instructions to be presented? -->
<!-- Please include detailed testing steps, explaining how to test your change. -->
<!-- Bear in mind that context you working on is not obvious for everyone.  -->
<!-- Adding "simple" configuration steps will help reviewers to get to your PR as quickly as possible. -->
<!-- "Before / After" screenshots can also be very helpful when the change is visual. -->

* In `projects/packages/blocks`, run `./vendor/bin/phpunit --filter Test_Blocks`
* In `projects/plugins/jetpack`, run `./vendor/bin/phpunit --filter WP_Test_Jetpack_Gutenberg`
* General block usage; can log out the names that are being registered
* Can check # of calls to `wp_json_file_decode()` in `wp-includes/functions.php` to verify reduced disk access